systemd-networkd: disable DAD for fallback IPv4 address
Our configuration with the static IPv4 fallback address suffers from
the following upstream regression in systemd v250:

https://github.com/systemd/systemd/issues/22763

This is a backport of the corresponding fix.

It also disables DAD for the fallback IP address,
otherwise it might be removed when a second device
is detected. While this should be re-considered later,
this restores the current/older behavior for now.

Signed-off-by: Michael Heimpold <michael.heimpold@in-tech.com>