bsp-network-config: drop IPMasquerade from eth2
This line triggers a warning message in Yocto kirkstone:
--snip--
IPMasquerade=yes is deprecated, and it is handled as "ipv4" instead
of "both". Please use "ipv4" or "both".
--snap--

However, this was broken by design - we do not need to masquerade
on the PLC link to upstream home/company network. The original reason for
doing was due to the fact, that QCA700x's firmware is only capable
to handle a single MAC address on the powerline.
Thus, it is not possible to bridge the eth2 interface with other
Ethernet interfaces, however, routed packets and connections
originating from the host itself always have the (source) MAC
address of the host - which is fine.

Signed-off-by: Michael Heimpold <michael.heimpold@in-tech.com>