lmsensors: add patches for upstream pwmfan driver support
The pwmfan driver in upstream Linux kernel comes with changed hwmon API.
But it implements the pwm1_enable file in a different way to
existing drivers.

As a workaround, we patch lmsensor's fancontrol and pwmconfig script
to check for this special driver and ignore the pwm1_enable
functionality so far.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>